sim: split up *ReplayRenderer classes into dedicated files.
