+++ src/tests/ReplayBatchRendererTest.cpp
@@ -13,7 +13,8 @@
 #include "esp/physics/objectManagers/RigidObjectManager.h"
 #include "esp/sensor/CameraSensor.h"
 #include "esp/sensor/Sensor.h"
-#include "esp/sim/ReplayBatchRenderer.h"
+#include "esp/sim/BatchReplayRenderer.h"
+#include "esp/sim/ClassicReplayRenderer.h"
 #include "esp/sim/Simulator.h"
 
 #include <Corrade/TestSuite/Compare/Numeric.h>
@@ -73,7 +74,7 @@ const struct {
      }},
     {"batch renderer", [](const ReplayRendererConfiguration& configuration) {
        return Cr::Containers::Pointer<esp::sim::AbstractReplayRenderer>{
-           new esp::sim::ReplayBatchRenderer{configuration}};
+           new esp::sim::BatchReplayRenderer{configuration}};
      }}};
 
 ReplayBatchRendererTest::ReplayBatchRendererTest() {
